Extract function and add basic tests
